Got main scene working with restructured engine. Added file reader/writer class. Shaders now load from CSO files, rather than compiling at runtime. Fixed jankiness from having separate timers in the Window class. Numerous fixes for cleaning up on exit. Fixed heat haze/blizzard effects.

+++ DirectX 11 Engine/Core/Graphics/UI/UserInterface.h
@@ -28,6 +28,6 @@ public:
 
 	void AddBitmapToUI(XMFLOAT2 bitmapSize, XMFLOAT2 bitmapPos, ID3D11ShaderResourceView* bitmapTexture);
 
-	void Update(float delta);
+	void Update(double delta);
 	void Draw();
 };
